requirements.txt
python-dotenv
aiogram == 2.5
langchain-google-genai
google-generativeai
IPython
chromadb
PyPDF2
python-docx
filetype
docx2txt
langchain-community
langchain-google-vertexai
google-auth
google-auth-oauthlib 
google-auth-httplib2 
google-cloud-aiplatform
matplotlib 
seaborn
requests 
pandas 
openpyxl 
xlsxwriter